otel/python/requirements.txt
fastapi[standard]
uvicorn[standard]
#opentelemetry-api
opentelemetry-sdk
opentelemetry-distro
# opentelemetry-instrumentation-fastapi
#opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http
